+++ ActiveField.php
@@ -7,7 +7,6 @@
 
 namespace yii\bootstrap;
 
-use yii\helpers\Html;
 use yii\helpers\ArrayHelper;
 
 /**
@@ -287,6 +286,22 @@ class ActiveField extends \yii\widgets\ActiveField
             };
         }
         parent::radioList($items, $options);
+        return $this;
+    }
+
+    /**
+     * Renders Bootstrap static form control.
+     * @param array $options the tag options in terms of name-value pairs. These will be rendered as
+     * the attributes of the resulting tag. There are also a special options:
+     *
+     * - encode: boolean, whether value should be HTML-encoded or not.
+     *
+     * @return $this the field object itself
+     */
+    public function staticControl($options = [])
+    {
+        $this->adjustLabelFor($options);
+        $this->parts['{input}'] = Html::activeStaticControl($this->model, $this->attribute, $options);
         return $this;
     }